requirements.txt
# PDF文本提取工具依赖

# 文字模式提取
PyPDF2>=3.0.0

# OCR模式提取
pdf2image>=1.16.0
pytesseract>=0.3.10
Pillow>=10.0.0
